pyproject.toml
[project]
name = "project-nexus"
version = "0.1.0"
description = "Multi-tenant AI agent framework"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.3.0",
  "httpx>=0.27.0",
  "pyyaml>=6.0.2",
  "redis>=5.0.7",
  "supabase>=2.6.0",
  "sqlalchemy>=2.0.31",
  "psycopg[binary]>=3.2.1",
  "sentence-transformers>=3.0.1",
  "tiktoken>=0.7.0",
  "langfuse>=2.39.0",
  "faker>=27.0.0",
  "aiogram>=3.10.0",
  "orjson>=3.10.6",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
voice = [
  "edge-tts>=6.1.0",
  "openai-whisper>=20240930"
]
integrations = [
  "mcp>=1.0.0",
  "structlog>=24.0.0"
]
dev = [
  "pytest>=8.3.2",
  "pytest-asyncio>=0.23.8",
  "ruff>=0.5.6",
  "mypy>=1.11.1"
]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
